Refactor several post_spec methods into its own user_spec.rb file
