+++ spec/features/posts_spec.rb
@@ -18,33 +18,5 @@ RSpec.describe Post, type: :feature do
       expect(page).to have_content 'This is quality content'
       expect(page).to have_content 'More quality content'
     end
-
-    it 'does send friend requests' do
-      sign_in users(:arnold2)
-      visit user_path(id: users(:juandavid).id)
-      click_button 'Send friend request'
-      expect(page).to have_content 'Friend request sent'
-    end
-
-    it 'does see friend requests' do
-      sign_in users(:juan)
-      visit '/friend_requests'
-      expect(page).to have_content 'Juan David'
-      expect(page).to have_content 'Pending'
-    end
-
-    it 'does accept a friend request' do
-      sign_in users(:juandavid)
-      visit '/friend_responses'
-      click_link 'Accept'
-      expect(page).to_not have_content 'Pending'
-    end
-
-    it 'does reject a friend request' do
-      sign_in users(:juandavid)
-      visit '/friend_responses'
-      click_link 'Reject'
-      expect(page).to have_content 'Rejected'
-    end
   end
 end
